refactor: move over the profile files into profile module
Move the files over to the profile module, there are still some issues running tests as the files ar
e not shared, so a test module will be made for this
