+++ app-profile/build.gradle
@@ -23,6 +23,17 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+    sourceSets {
+        androidTest {
+            java.srcDirs += "../core/src/sharedTest/java"
+        }
+        test {
+            java.srcDirs += "../core/src/sharedTest/java"
+        }
+    }
+    dataBinding {
+        enabled true
+    }
     androidExtensions {
         experimental true
     }
@@ -34,19 +45,32 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
+    implementation project(path: ':account')
+    implementation project(path: ':core')
+    implementation project(path: ':series')
+
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
     implementation "androidx.appcompat:appcompat:$appcompat_version"
+    implementation "androidx.constraintlayout:constraintlayout:$constraint_layout_version"
     implementation "androidx.core:core-ktx:$corektx_version"
+    implementation "androidx.lifecycle:lifecycle-extensions:$lifecycle_version"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
     implementation "com.chesire:lifecyklelog:$lifecykle_version"
+    implementation "com.github.bumptech.glide:glide:$glide_version"
     implementation "com.google.dagger:dagger:$dagger_version"
     implementation "com.google.dagger:dagger-android:$dagger_version"
     implementation "com.google.dagger:dagger-android-support:$dagger_version"
+    implementation "com.google.android.material:material:$material_version"
 
+    testImplementation 'androidx.arch.core:core-testing:2.1.0-rc01'
+    testImplementation 'io.mockk:mockk:1.9.3'
     testImplementation 'junit:junit:4.12'
+    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:$coroutines_version"
 
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
 
+    kapt "com.github.bumptech.glide:compiler:$glide_version"
     kapt "com.google.dagger:dagger-compiler:$dagger_version"
     kapt "com.google.dagger:dagger-android-processor:$dagger_version"
 }
