+++ app-profile/src/main/java/com/chesire/malime/app/profile/ProfileFragment.kt
@@ -1,28 +1,20 @@
-package com.chesire.malime.flow.profile
+package com.chesire.malime.app.profile
 
 import android.os.Bundle
 import android.view.LayoutInflater
-import android.view.Menu
-import android.view.MenuInflater
-import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.get
-import androidx.navigation.fragment.findNavController
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import com.chesire.lifecyklelog.LogLifecykle
-import com.chesire.malime.R
+import com.chesire.malime.app.profile.databinding.FragmentProfileBinding
 import com.chesire.malime.core.models.UserModel
 import com.chesire.malime.core.viewmodel.ViewModelFactory
-import com.chesire.malime.databinding.FragmentProfileBinding
 import dagger.android.support.DaggerFragment
-import kotlinx.android.synthetic.main.fragment_profile.fragmentProfileAnimeProgress
 import kotlinx.android.synthetic.main.fragment_profile.fragmentProfileAvatar
-import kotlinx.android.synthetic.main.fragment_profile.fragmentProfileMangaProgress
-import kotlinx.android.synthetic.main.view_series_progress.view.viewSeriesProgressTitle
 import javax.inject.Inject
 
 @LogLifecykle
@@ -62,18 +54,6 @@ class ProfileFragment : DaggerFragment() {
                 setImagery(it)
             }
         })
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        super.onCreateOptionsMenu(menu, inflater)
-        inflater.inflate(R.menu.menu_profile, menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        when (item.itemId) {
-            R.id.menuProfileSettings -> findNavController().navigate(ProfileFragmentDirections.toSettings())
-        }
-        return super.onOptionsItemSelected(item)
     }
 
     private fun setImagery(user: UserModel) {
