+++ app-profile/src/main/res/layout/fragment_profile.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="vm"
-            type="com.chesire.malime.flow.profile.ProfileViewModel" />
+            type="com.chesire.malime.app.profile.ProfileViewModel" />
     </data>
 
     <androidx.coordinatorlayout.widget.CoordinatorLayout
