+++ app-profile/src/main/res/layout/view_series_progress.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="seriesProgress"
-            type="com.chesire.malime.flow.profile.SeriesProgress" />
+            type="com.chesire.malime.app.profile.SeriesProgress" />
     </data>
 
     <com.google.android.material.card.MaterialCardView
