--- app/src/main/res/menu/menu_profile.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-    <item
-        android:id="@+id/menuProfileSettings"
-        android:icon="@drawable/ic_settings_white"
-        android:title="@string/menu_settings"
-        app:showAsAction="ifRoom" />
-</menu>
